Remove docs directory from make file
The docs directory does not contain any files at all and should because
of that not be included in the build process. Incidently this empty
directory does not trigger any errors except for "make dist" - part of
the process to make actual releases.

Thanks to Klaus Knopper who pointed out the solution to this.