[#3572032]Removed preliminary pushing requests into ready_queue.
It might create conflict in case the request was processed in ready_queue
and therefore freed but still exists in task. Such conflict may happen if
there is cached thumbnail of one of sizes while there is request for both
large and small sizes. To avoid it we'll delay request return to time when
both sizes are generated. Such case is anyway rare enough and moreover,
the only effect it may cause is a little delay for thumbnail of other size
so both will be displayed at once. This is acceptable price for crash-free
implementation.