Don't set any defaults on context menu sorting.
If we set some defaults on context menu sorting then we get them into
application callback when folder view is created so if model wasn't set
at that moment this may reset application defaults into common defaults.
Application may test whether own sort was set but that isn't right.
This is a bug of FmFolderView -- not FmFolderView should set sorting
but application; therefore that setting was removed.