[#3584608]Fix for rarely broken DnD due to FmDndDest source caching.
The previous behavior was to cache data and reset cache when drag leaves
widget. This is completely wrong - we never get informed if we refuse drop
(for example, on wrong target) therefore drag source kept the same while
it was already changed. The only valid behavior is to keep drag context
referenced and compare it with one we get in handler and if it's changed
then reset our cache. This way we don't have to requery data again when
pointer leaves widget but still inside of application window - the drag
context is still the same, i.e. we get some optimization as a side effect.