Removing workaround on http://bugzilla.xfce.org/show_bug.cgi?id=6230
The workaround isn't required anymore since it was fixed in libexo
after the workaround was added and that fix is present in the libfm
source file src/gtk/exo/exo-tree-view.c already.

P.S. It is proven to be fixed - checked on both gtk 2.24 and 2.20.1.
The 2.20.1 is the one where the bug was experinced the first time
and that was on Ubuntu Lucid 10.04, so checked exactly on the Lucid.