Implementing ask user to mount unmounted path.
If we try to get info on path which isn't mounted
there is reason to ask user to mount it.
That was dropped in 7d9c1295a3d785122a3d7a141d54698f3c5120dc due to
inappropriate code placement and now we can handle it due to ABI change.