smb:// symlinked locations cannot be launched from folder view.
This is because we don't assume each symlink is a folder anymore
(that assuming was stupid anyway) so we check how to handle the
symlink target. We have to get info on the target. If target isn't
mounted we cannot retrieve info on it and FmFileInfoJob fails.
Therefore we cannot launch the location.
The commit adds creation of simplest file info which contains only
name and type (directory) for not mounted location.