Make action 'NewBlank' to work again via fm_folder_view_bounce_action() API.
The action 'NewBlank' is bounced by pcmanfm despite it was never
documented and omitting it will drop some functionality.
Since there is no such special behavior in templates (each template
defines some kind of file) we cannot support it via template.
This commit reimplements such behavior for fm_folder_view_bounce_action()
call. It also documents this action in description of this function.
This action isn't present in context menu list as it may conflict
with some of templates (unlikely but it is still possible).