Fix: file owner/group should be properly disabled in dialog.
Currently if owner and group cannot be changed by user those fields
are enabled for set cursor but content is unchangeable. That may
be questioned by users and considered as bug, it also visually not
distinguishable from changeable case.
The commit properly disables those fields completely so user will
see it cannot be changed and cannot set keyboard focus there.