Adding tooltips for file name and location in file properties.
In case when file name is too long to be shown in the field it will be
truncated so user cannot see it. Adding tooltip with full file name
and full file path helps to see it in any case.
The commit implements showing tooltips on base of text lenght, not
rendering size. That probably should be changed later.