Change FM_CONFIG_DEFAULT_NO_USB_TRASH to more expecting TRUE.
The most expected for users used to Windows behavior for that option
is to have it TRUE, and FALSE creates behavior unexpected by most of
the users and, moreover, even when user plug off flash drive an then
plug in it, the trash may be not always recognized by GVFS therefore
it's not trivial to get files on flash drive be REALLY deleted. And
users blame pcmanfm, of course.
The commit changes default. It will not affect users who already use
libfm 1.1.0 and have own setting for it but only new users.