Support for columns width in Detailed List View mode.
The commit adds support for Detailed List View mode in FmStandardView:
- added a 'width' member into FmFolderViewColumnInfo struct; value 0
  means auto-size; value >0 means fixed width.
- the 'width' is applied on column setup.
- when user manually changes any column except for 'Name' the 'width'
  is updated accordingly.

Column 'Name' is a special one - it is expandable so it will have width
depending on window size. Minimum size is 200 pixels. That value can be
changed by user but I could not find any reliable way to distinguish
that manual change from window resize or folder view update on redraw,
therefore that change isn't detectable for now.