Correction on spacing in file properties when file isn't symlink.
In case when file is symlink the target file is shown in General tab of
file properties dialog. In case when file isn't symlink it is not shown
but extra reserved space is still there therefore spacing in that place
is bigger than it should be. This commit fixes that.