Use FmXmlFile to parse .menu files and correctly change Categories on item move.
The former implementation was very much incorrect since it tried to change
Categories with respect to last past member but unfortunately it is the same
only for few categories. Real relations of Categories and path is much more
complicated (it may require some special combinations to work).
With help of FmXmlFile parser we can create merged menu representation and
analyze it, and finally edit Categories in desktop entry to do real movement.
This commit adds working implementation for applications moving. It doesn't
implement rdding/deleting/enaming of directories yet.