Implementation: add and delete folders under menu://applications/
Since we have FmXmlFile container, we can edit user menu file to do things.
There is only problem with that: libmenu-cache before 0.5.0 does not support
'show hidden' for directories and empty folder cannot be seen even with the
'show hidden' enabled so user cannot see newly created empty directory, it
will become visible if you manually edit some desktop entry to enable the
Category for it. But if there were some desktop entries for it then folder
will be visible immediately. So using it may be problematic for inexperienced
users. Probably it is better to disable the feature if compiled with any
version of libmenu-cache prior to 0.5.0.