Separate library for non-FM-related classes such as FmXmlFile.
Some classes can be used by programs that don't use any folder related functions
but are still useful. Example is FmXmlFile (and also FmModules that is planned
for implementation). In some cases it is even impossible to use them in libfm,
just for example, menu cache generator in libmenu-cache cannot relate on libfm
which is dependent on it, that will create a cyclic dependency, but FmXmlFile
can be useful for that generator.
A solution for that is to create a library which have only GLib as dependency.
That library can be used by main (i.e. libfm) and still be around but it can be
used as standalone in the same time. I named it libfm-extra, see the commit.