Fix for Gtk-Critical on accel_group on window destroy.
Sometimes we get Gtk-Critical on window close. The cause is the logic when
detaching GtkAccelGroup from window is based on locked state of the group.
That logic (attempt to simplify behavior) was wrong, we should check if the
accel group is present on window and try to remove it only if it is there.