Show "folder" icon if directory's icon cannot be loaded.
There is a case when icon for some directory in Applications subfolders
cannot be loaded and "unknown" icon is shows. There is no reason to
show it as it will be not distiguishable from unknown file. With help
of fm_pixbuf_from_icon_with_fallback() we now can show "folder" icon
correctly on such folders so they will look different from regular files.