Wrong arg to fm_folder_model_get_column_type() should return value not assertion.
The g_return_val_if_fail() is an assertion which works only in debug mode
but it will do nothing in normal mode. And receive incorrect values in
some cases should be made possible instead of crashing application.
The commit changes it to plain if() for fm_folder_model_iter_n_children()
and fm_path_entry_changed() too.