Set busy cursor on new created window in libfm-demo correctly.
It was set when main window wasn't realized yet therefore failed and
it was never possible to see if folder is still loading or already loaded.

If we set default size and realize window before we start loading the
folder then it will work correctly. Commit adds this into fm_main_win_init().