[#3578780]Activating item is ambiguos when there are more than one item selected.
Let do it the same way as other file managers do it, i.e.
unselect all other items first and then activate only last active one.
This will work in both cases:
- if user (double)clicks the item;
- if user navigates from keyboard then presses 'Enter'.

It works a bit odd if user selected items by mouse and then press 'Enter': it
will activate last active item as well but ExoIconView doesn't take the range
selection into account for that. May be that is a bug but may be not.