Item LinkTo belongs to PCManFM and SendTo is rather a category.
The Link item is similar to CopyTo or MoveTo and therefore better to be in PCManFM.
The "Send to" is a category which is already present in meny as AddBookmark or
Compress. It is better to just allow extensions to add items there instead.
Since both those items were never implemented, this commit just removes unused text.