[#3614848]Allow search in URIs, not only in native paths.
Implementation of search does not allow neither ':' nor double slashes
in search path. Also there is no escaping for '&' in search pattern.
To support this we should escape all mentioned elements before they are
concatenated into search:// URI, we cannot escape them after they are
already added because we cannot distinguish separator from patterns.
Therefore escaping should be done in the widget but unescaping should
be done for each element separately, due to the same reason - we cannot
distinguish them after they are unescaped. Escaping should be done with
standard for URI way, it is the most safe way for GLib normalizations.

The commit adds those escapings and unescapings.