Append user's modules_blacklist to system one, not forget system.
System administrator may want to set blacklist for some unwanted modules
so user should not drop it. User's blacklist is appended to system,
the system length is remembered and user's blacklist is still saved.
Note that application should not replace modules_blacklist in the config
otherwise it will be saved incorrectly.

This approach is dirty but I see no other right now - either we don't
allow application to edit the list, or we don't show system blacklist
to application at all. What is better?