Use XML file for menu editing, don't edit .desktop files.
Despite the fact categories manipulations is working well, it will leave
user entry as it was last time it was moved despite it is the same system
entry and if system entry will change 'Exec' or something else, user entry
will not change. Editing XML menu is a bit better in this regard.
It does affects only current desktop environment though but that should be
not much a problem for the most of users.