[#3614500]Menu should be detached if parent window is destroyed.
It is probably cause of bug. It is possible in theory that top window
is destroyed while nested widget is still not finalized (for example
due to references that hold it from finalize) and therefore attached
widget on menu becomes invalid before menu is destroyed. We should call
gtk_menu_detach() in such case. Let see if this commit fixes the issue.