Don't show 'Trash' in context menu on non-native files but 'Delete' instead.
In many cases trashing files on virtual filesystems isn't supported
and this may confuse users. If user still wants to try trash such files,
still drag&drop and cut&paste are available. But if trashing is enabled
for native files only we avoiding error window "files cannot be trashed"
and we don't have to check for trash URI as well.