Fix handling for shortcuts to desktop entries.
Shortcut to desktop entry is very special case: it should be detected
as inode/x-shortcut still so it will be not edited as desktop entry, it
will fail in most of cases. And launching such shortcuts should also be
handled not via default MIME type detection but by special way, the same
as we do for desktop entries, that special case wasn't tested yet.