Move standard columns initialization into _fm_folder_model_init().
In case it is in class initialization we not only fail on column verification
before FmFolderModel class is initialized but also we may get a race condition
from loaded modules since they will try to reallocate not allocated yet data,
and if class init is later that init will abandon already loaded data.
Therefore we should do init of columns array before we schedule modules load.