[#3615226]Remove item 'Rename' from folder context menu, it's confusing.
If we rename the folder itself then we get only G_FILE_MONITOR_EVENT_DELETED
notification and therefore folder view is closed. Add some dirty hacks on the
event is impossible too because we should add references to API therefore
make it incompatible. So only choice is to remove the item from menu.