Make drop files on Desktop more "smart": copy files from plugged media.
It is still preferrable for inexperienced users to not move ~/Movies into
~/Desktop accidentally and to not save html code when dropping URL but files
from plugged in USB stick should be copied instead. Therefore we should
test for that case and change default auto-action. Of course, it will be
not applied if user not selected 'auto' as 'drop_default_action' or if user
pressed 'Shift' or 'Ctrl' when dropped files.