Don't assume shortcuts are directories on launch - it is wrong.
Previous commit is wrong in that assumption. In most of cases a shortcut is
not a directory and if we assume shortcut is a directory then we never can
launch any shortcut to some application, HTML link, or whatever else.

Also if a shortcut is a directory then that will be handled there by
fm_file_info_is_dir(fi) condition anyway.