Prevent corrupted image thumbnails
Since `generate_thumbnails_with_builtin()` is in thread, it may not have time to return a TRUE value in `fm-thumbnail-loader.c → generate_thumbnails()`, so that if there is an external image thumbnailer (like the PNG thumbnailer of gdk-pixbuf2 >= 2.36.1), a corrupted thumbnail will be imminent.

Please see that the above statement is just a speculation. However, this patch really prevents creation of corrupted thumbnails of PNG images with gdk-pixbuf2 >= 2.36.1. I also did some tests with a `gboolean`, to which the return value of `generate_thumbnails_with_builtin()` was assigned, and it always remained equal to its initial value, whether it was TRUE or FALSE.