Add a setting to open files with a single middle-click
With this setting, a single middle-click has the same effect as a
double-click with the left mouse button.

The single left-click and the single middle-click settings are mutually
exclusive, i.e. both settings cannot be active at the same time.

This uses the middle mouse button differently to pull request #46
by livanh.